solving out of memory error on brain_large dataset

+++ scvi/train.py
@@ -11,7 +11,7 @@ def train(vae, data_loader_train, data_loader_test, n_epochs=20, learning_rate=0
     # Training the model
     for epoch in range(n_epochs):
         for i_batch, (sample_batch, local_l_mean, local_l_var, batch_index, _) in enumerate(data_loader_train):
-
+            sample_batch = sample_batch.type(torch.FloatTensor)
             sample_batch = Variable(sample_batch)
             local_l_mean = Variable(local_l_mean)
             local_l_var = Variable(local_l_var)
